news_feed/backend/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
django-environ==0.9.0
djoser==2.1.0
drf-yasg==1.21.4
gunicorn==20.1.0
ipython==8.7.0
psycopg2-binary==2.8.6
sort-requirements==1.3.0
